[BM_OSAL/HAL] Decoupled WD completely from ST LL and FreeRTOS
