+++ src/lib/CMakeLists.txt
@@ -1,3 +1,5 @@
+set(LIB_DIR ${CMAKE_CURRENT_LIST_DIR})
+
 # Include TinyUSB
 add_subdirectory(${LIB_DIR}/usb)
