+++ src/lib/bm_osal/bm_osal.h
@@ -6,6 +6,11 @@
 #include <stdint.h>
 #include <stdbool.h>
 
+// Define a base return type for RTOS APIs
+typedef int bm_osal_base_t;
+
+typedef int bm_osal_tick_type_t;
+
 // Mutex
 typedef void* bm_osal_mutex_t;
 bool bm_osal_mutex_create(bm_osal_mutex_t* mutex);
@@ -25,6 +30,8 @@ typedef void * bm_task_handle_t;
 typedef void (*bm_osal_task_func_t)(void* arg);
 bool bm_osal_task_create(bm_osal_task_func_t task_func, const char* name, uint16_t stack_size, void* arg,  uint32_t priority, bm_task_handle_t task_handle);
 bool bm_osal_task_delete(void);
+uint16_t bm_osal_task_get_min_stack_size();
+void bm_osal_task_delay(bm_osal_tick_type_t delay);
 
 // Timer
 typedef void* bm_osal_timer_t;
@@ -36,5 +43,10 @@ bool bm_osal_timer_delete(bm_osal_timer_t timer);
 
 // Delay
 void bm_osal_delay_ms(uint32_t ms);
+
+// void bm_hal_wd_reload_counter(void* handle);
+void bm_hal_wd_reload_counter(void);
+// void* bm_hal_wd_get_handle(void);
+void bm_hal_wd_log();
 
 #endif // BM_OSAL_H
